Programação da ficha : faltando a ordenação da lista ,quando houver a remoção de uma serie , de um exercicio e suas series por completo e a movimentação da serie ! metodos para reordenar na TreinoDAO,ControleTreino e GUIFichaTreino

+++ WorkOutSystem/src/workoutsystem/view/GUIFichaManipular.java
@@ -442,7 +442,7 @@ ListView.OnItemLongClickListener{
 				mensagem = controleTreino.manipularTreino
 				(editNomeTreino.getText().toString(),
 						ficha.getCodigo(),
-						Long.parseLong(txtCodigoTreino.getText().toString()));
+						(int) Long.parseLong(txtCodigoTreino.getText().toString()));
 				ficha = controleFicha.buscarFichaCodigo(ficha.getCodigo());
 				createListView(ficha.getTreinos());
 				dialog.dismiss();
